Add property scan
