+++ src/include/planner/join_order/cardinality_estimator.h
@@ -42,6 +42,9 @@ private:
     uint64_t atLeastOne(uint64_t x) { return x == 0 ? 1 : x; }
 
     uint64_t getNodeIDDom(const std::string& nodeIDName) {
+        if (!nodeIDName2dom.contains(nodeIDName)) {
+            return 100;
+        }
         KU_ASSERT(nodeIDName2dom.contains(nodeIDName));
         return nodeIDName2dom.at(nodeIDName);
     }
