+++ src/planner/plan/plan_read.cpp
@@ -8,6 +8,8 @@
 #include "planner/operator/scan/logical_scan_node_table.h"
 #include "planner/operator/sip/logical_property_collector.h"
 #include "planner/planner.h"
+#include "planner/operator/logical_table_function_call.h"
+#include "function/table/hnsw/hnsw_index_functions.h"
 
 using namespace kuzu::binder;
 using namespace kuzu::common;
@@ -118,9 +120,21 @@ void Planner::planTableFunctionCall(const BoundReadingClause& readingClause,
     splitPredicates(call.getColumns(), call.getConjunctivePredicates(), predicatesToPull,
         predicatesToPush);
     for (auto& plan : plans) {
-        planReadOp(getTableFunctionCall(readingClause), predicatesToPush, *plan);
+        auto op = getTableFunctionCall(readingClause);
+        op->computeFactorizedSchema();
+        planReadOp(op, predicatesToPush, *plan);
         if (!predicatesToPull.empty()) {
             appendFilters(predicatesToPull, *plan);
+        }
+        auto callOp = op->ptrCast<LogicalTableFunctionCall>();
+        if (StringUtils::getUpper(callOp->getTableFunc().name) == function::QueryHNSWIndexFunction::name) {
+            auto bindData = callOp->getBindData()->constPtrCast<function::QueryHNSWIndexBindData>();
+            auto node = bindData->outputNode;
+            auto scanPlan = LogicalPlan();
+            appendScanNodeTable(node->getInternalID(), node->getTableIDs(), node->getPropertyExprs(), scanPlan);
+            expression_vector joinConditions;
+            joinConditions.push_back(node->getInternalID());
+            appendHashJoin(joinConditions, JoinType::INNER, *plan, scanPlan, *plan);
         }
     }
 }
@@ -161,6 +175,8 @@ void Planner::planGDSCall(const BoundReadingClause& readingClause,
         }
     } else {
         for (auto& plan : plans) {
+            auto gdsCall = getGDSCall(call.getInfo());
+            gdsCall->computeFactorizedSchema();
             planReadOp(getGDSCall(call.getInfo()), predicatesToPush, *plan);
         }
     }
@@ -197,6 +213,7 @@ void Planner::planLoadFrom(const BoundReadingClause& readingClause,
         predicatesToPull, predicatesToPush);
     for (auto& plan : plans) {
         auto op = getTableFunctionCall(*loadFrom.getInfo());
+        op->computeFactorizedSchema();
         planReadOp(std::move(op), predicatesToPush, *plan);
         if (!predicatesToPull.empty()) {
             appendFilters(predicatesToPull, *plan);
@@ -206,7 +223,6 @@ void Planner::planLoadFrom(const BoundReadingClause& readingClause,
 
 void Planner::planReadOp(std::shared_ptr<LogicalOperator> op, const expression_vector& predicates,
     LogicalPlan& plan) {
-    op->computeFactorizedSchema();
     if (!plan.isEmpty()) {
         auto tmpPlan = LogicalPlan();
         tmpPlan.setLastOperator(std::move(op));
