+++ src/processor/map/map_copy_from.cpp
@@ -1,6 +1,6 @@
 #include "catalog/catalog_entry/node_table_catalog_entry.h"
 #include "catalog/catalog_entry/rel_table_catalog_entry.h"
-#include "function/table/call/create_hnsw_index.h"
+#include "function/table/hnsw/hnsw_index_functions.h"
 #include "planner/operator/logical_partitioner.h"
 #include "planner/operator/persistent/logical_copy_from.h"
 #include "processor/expression_mapper.h"
